refactor(test): move test files to __tests__ directory ``` </think>
refactor(test): move test files to __tests__ directory
